requirements.txt
gitdb==4.0.9
GitPython==3.1.27
smmap==5.0.0
typing-extensions==4.3.0